Move scream and moan sounds to medical feedback
Also moving the vanilla feedback changes over
