+++ addons/medical_engine/XEH_PREP.hpp
@@ -1,6 +1,4 @@
-
 PREP(handleDamage);
-PREP(playInjuredSound);
 PREP(damageBodyPart);
 PREP(updateBodyPartVisuals);
 PREP(setLimping);
