+++ addons/medical_feedback/functions/fnc_playInjuredSound.sqf
